Updated JUnit for q1 and added test for output from the main method
